requirements.txt
absl-py
sentencepiece
tensorflow>=2.0.0
tensorflow_text>=2.0.0
